Use isKeyDown for key binds to prevent odd issues during movement.
If isKeyDown is not used, bugs would occur such as constant jumping while
moving forward. Before isKeyDown is called, a check is performed to make
sure the bind set is not on a mouse. If isKeyDown gets called with a mouse
bind, the game crashes.

* Allow horse breeding between Unicorn and Black Pegasus. This will result
in a Black Fairy.
* Fix normal undead skeletons not reverting to zombie form when fed essence of
light.
* Fix ostriches always being tamed.